+++ app/src/main/java/com/example/calculadora_reviso/MainActivity.kt
@@ -6,7 +6,7 @@ import android.widget.Toast
 import com.example.calculadora_reviso.databinding.ActivityMainBinding
 import java.text.DecimalFormat
 
-
+//faz a exibição do código
 class MainActivity : AppCompatActivity() {
     lateinit var binding: ActivityMainBinding
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -21,26 +21,26 @@ class MainActivity : AppCompatActivity() {
         }
     }
 
-               fun calculate(){
-                  val InsertEdit = binding.Insert //edittext
-                  val Insert2Edit = binding.Insert2//editText
-                  val selectedid = binding.radio.checkedRadioButtonId//RadioButton selecionado
+       fun calculate(){
+          val InsertEdit = binding.Insert //edittext
+          val Insert2Edit = binding.Insert2//editText
+          val selectedid = binding.radio.checkedRadioButtonId//RadioButton selecionado
 
-                  //convertendo variaveis
+          //convertendo variaveis
 
-                  val N1 = InsertEdit.text.toString().toInt() //conveti as variavies para int
-                  val N2 = Insert2Edit.text.toString().toInt() //conveti as variavies para int
+          val N1 = InsertEdit.text.toString().toFloat() //conveti as variavies para float
+          val N2 = Insert2Edit.text.toString().toFloat() //conveti as variavies para float
 
-                   //criei uma variavel com uma estrutura de decisao, se um radiobutton for selecionado executa
-                   //uma funçao, com os parametros das variaveis (N1 e N2)
-                   val tipcalculation = when (selectedid){
-                       R.id.add -> Calculadora().somar(N1, N2)
-                       R.id.multiply -> Calculadora().multiplicar(N1, N2)
-                       R.id.subtract -> Calculadora().menos(N1, N2)
-                       else -> Calculadora().divisao(N1, N2)
-                   }
+           //criei uma variavel com uma estrutura de decisao, se um radiobutton for selecionado executa
+           //uma funçao, com os parametros das variaveis (N1 e N2)
+           val tipcalculation = when (selectedid){
+               R.id.add -> Calculadora().somar(N1, N2)
+               R.id.multiply -> Calculadora().multiplicar(N1, N2)
+               R.id.subtract -> Calculadora().menos(N1, N2)
+               else -> Calculadora().divisao(N1, N2)
+           }
 
-                   val tipResult = DecimalFormat().format(tipcalculation)
-                   binding.result.text = tipResult
-              }
+           val tipResult = DecimalFormat().format(tipcalculation)
+           binding.result.text = tipResult
+      }
 }
